+++ R/filter.R
@@ -6,29 +6,29 @@
 #'
 #' @param .data \linkS4class{AcousticStudy} or \linkS4class{AcousticEvent} to filter
 #' @param \dots Logical expressions, syntax is identical to \link[dplyr]{filter}.
-#'   There are special cases to filter by environmental variables, species ID, 
+#'   There are special cases to filter by environmental variables, species ID,
 #'   database, or detector name. See details.
 #' @param .preserve not used
 #'
 #' @details Most expression provided will be used to filter out detections based on
-#'   calculated parameters. 
-#'   
+#'   calculated parameters.
+#'
 #'   If the name of an environmental variable added using
 #'   \link{matchEnvData} is provided, will filter to only events with environmental
-#'   variables matching those conditions. 
-#'   
+#'   variables matching those conditions.
+#'
 #'   If a provided logical expression uses
 #'   \code{"species"} or \code{"Species"}, then events will be filtered using the
 #'   species present in the \code{$id} of the \code{species} slot of each event.
-#'   
-#'   If a provided logical expression uses \code{"database"} or \code{"Database"}, 
+#'
+#'   If a provided logical expression uses \code{"database"} or \code{"Database"},
 #'   then only events with databases matching the expression in \code{files(.data)$db}
 #'   will remain
-#'   
+#'
 #'   If a provided logical expression uses \code{"detector"} or \code{"Detector"}, then
 #'   only detections from detectors with names matching the expression will remain in
 #'   events. Any events left with no detections will be removed.
-#' 
+#'
 #' @return The original \code{.data} object, filtered by the given logical expressions
 #'
 #' @examples
@@ -148,18 +148,18 @@ filter.AcousticEvent <- function(.data, ..., .preserve=FALSE) {
             thisKeep[is.na(thisKeep)] <- FALSE
             detKeep <- detKeep & thisKeep
         }
-        if(!any(detKeep)) {
-            return(NULL)
-        }
+        # if(!any(detKeep)) {
+        #     return(NULL)
+        # }
         detectors(.data) <- detectors(.data)[detKeep]
     }
     detectors(.data) <- lapply(detectors(.data), function(x) {
         doFilter(x, ...)
     })
     detNums <- sapply(detectors(.data), nrow)
-    if(all(detNums == 0)) {
-        return(NULL)
-    }
+    # if(all(detNums == 0)) {
+    #     return(NULL)
+    # }
     detectors(.data) <- detectors(.data)[detNums > 0]
     .data
 }